requirements.txt
# Core ML/DL Framework
torch==2.2.2
torchvision==0.17.2

# Transformers and NLP
transformers==4.57.1
tokenizers==0.22.1
accelerate==1.11.0
timm==1.0.21
tiktoken==0.8.0
sentencepiece==0.2.0
nltk==3.9.2
langdetect==1.0.9

# LangChain Ecosystem
langchain==0.3.3
langchain-core==0.3.12
langchain-community==0.3.2
langchain-chroma==0.1.4
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langchain-unstructured==0.1.6
langsmith==0.1.147

# Vector Database
chromadb==0.5.20
chroma-hnswlib==0.7.6

# Document Processing
unstructured==0.18.15
unstructured-client==0.28.1
unstructured-inference==1.0.5
unstructured-pytesseract==0.3.15
python-docx==1.2.0
python-oxmsg==0.0.2
pypdf==6.1.3
pypdfium2==5.0.0
pymupdf==1.26.5
pikepdf==7.2.0
pdf2image==1.17.0
pdfminer.six==20250506
python-magic==0.4.27
filetype==1.2.0

# OCR
pytesseract==0.3.13
pillow==12.0.0
pillow-heif==0.18.0
pi-heif==1.1.1

# Image Processing
effdet==0.4.1
pycocotools==2.0.10

# Computer Vision Models
google-cloud-vision==3.11.0

# Web & API
fastapi==0.121.0
uvicorn==0.38.0
starlette==0.49.3
httpx==0.27.2
httpcore==1.0.9
httptools==0.7.1
h11==0.16.0
python-multipart==0.0.20
aiofiles==25.1.0
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
websockets==15.0.1
websocket-client==1.9.0
uvloop==0.22.1
watchfiles==1.1.1

# OpenAI
openai==1.40.6

# Data Processing
pandas==2.2.3
numpy==1.26.4
scipy==1.15.3

# Visualization
matplotlib==3.10.7
beautifulsoup4==4.14.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.3
click==8.3.0
typer==0.20.0
typer-slim==0.20.0
rich==14.2.0
tqdm==4.67.1
tabulate==0.9.0
coloredlogs==15.0.1
humanfriendly==10.0

# Pydantic
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.11.0

# Requests & HTTP
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
urllib3==2.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11

# Authentication & Crypto
cryptography==46.0.3
bcrypt==5.0.0
pycryptodome==3.23.0
oauthlib==3.3.1

# Storage & Serialization
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonpath-python==1.0.6
orjson==3.11.4
protobuf==6.33.0

# Database
sqlalchemy==2.0.44

# Telemetry
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0

# Google Cloud
google-api-core==2.28.1
google-auth==2.42.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0

# Kubernetes
kubernetes==34.1.0

# Deep Learning Optimization
onnx==1.19.1
onnxruntime==1.19.2
safetensors==0.6.2

# ML Utilities
joblib==1.5.2
ml-dtypes==0.5.3
sympy==1.14.0
mpmath==1.3.0

# Data Formats
lxml==6.0.2
html5lib==1.1
webencodings==0.5.1
olefile==0.47
omegaconf==2.3.0

# Compression
zstandard==0.25.0

# Hashing
mmh3==5.2.0

# Parsing
pyparsing==3.2.5
chardet==5.2.0
emoji==2.15.0

# Text Processing
regex==2025.11.3
rapidfuzz==3.14.3
markdown-it-py==4.0.0
mdurl==0.1.2
pygments==2.19.2

# Date/Time
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
python-iso639==2025.2.18

# Archives
flatbuffers==25.9.23

# Networking
backoff==2.2.1
posthog==5.4.0

# System
psutil==7.1.3
shellingham==1.5.4

# File System
fsspec==2025.10.0
hf-xet==1.2.0

# Build Tools
build==1.3.0
pyproject-hooks==1.2.0
tomli==2.3.0

# Google Auth
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
rsa==4.9.1

# Misc
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asgiref==3.10.0
async-timeout==4.0.3
attrs==25.4.0
cachetools==6.2.1
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
durationpy==0.10
eval-type-backport==0.2.2
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
greenlet==3.2.4
jinja2==3.1.6
jiter==0.11.1
kiwisolver==1.4.9
markupsafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
overrides==7.7.0
packaging==23.2
pycparser==2.23
proto-plus==1.26.1
propcache==0.4.1
pypika==0.48.9
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tenacity==8.5.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
mammoth==1.9.0